Update crhelper.py
If I am using this helper as intended there is a weird condition with Update requests that cause it not to fail correctly.
An exception thrown from a botched update causes the physical id to be changed to the lambda log stream name. Cloudformation sees the new physical resource id and assumes the update required a delete event to clean up the old physical resource id instead of just failing on the failure.